test: address javascript worker review blockers
